requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.5
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
dill==0.3.9
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.2
fastapi-cli==0.0.5
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
flower==2.0.1
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
humanize==4.11.0
idna==3.10
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
locust==2.32.4
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.1.0
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.1.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pylint==3.3.2
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.2.0
requests==2.32.3
rich==13.9.2
sentry-sdk==2.19.0
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.40.0
tomlkit==0.13.2
tornado==6.4.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
